+++ src/assets/data.js
@@ -1,0 +1,7 @@
+const skillsList = [
+  ['LANGUAGES: ', 'Ruby, JavaScrypt, HTML, CSS, MySQL'],
+  ['FRAMEWORKS: ', 'Ruby on Rails, React, Redux, Bootstrap, Sidekiq, Rspec, Capybara'],
+  ['TOOLS & METHODS: ', 'Github, Heroku, Netlify, Responsive design, TDD, Chrome Dev'],
+  ['SKILLS ', 'Pair-Programming, Problem-solving, Teamwork, Mentoring']];
+
+export default skillsList;
