+++ src/components/home/Skill.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import '../../CSS/home.css';
+import PropTypes from 'prop-types';
+
+const Skill = ({ tag, list }) => (
+  <p className="skillp text-center">
+    <span className="font-weight-bold">{tag}</span>
+    {list}
+  </p>
+);
+
+export default Skill;
+
+Skill.propTypes = {
+  tag: PropTypes.string.isRequired,
+  list: PropTypes.string.isRequired,
+};
